+++ lib/src/memory_cache_storage.dart
@@ -4,66 +4,74 @@
  */
 
 import 'package:resource_repository_storage/resource_repository_storage.dart';
+import 'package:rxdart/rxdart.dart';
 import 'package:synchronized/synchronized.dart';
 
 /// A simple implementation of [CacheStorage]. Used by default.
 /// It just keeps the data in memory.
-class SimpleMemoryCacheStorage<K, V> implements CacheStorage<K, V> {
+class MemoryCacheStorage<K, V> implements CacheStorage<K, V> {
   static final _lock = Lock();
-  static final Map<String, Map> _boxes = {};
+  static final Map<String, BehaviorSubject<Map>> _boxes = {};
 
   final String _boxKey;
 
   /// Create instance. Takes a unique key for storing data in separate boxes for every repository.
-  SimpleMemoryCacheStorage(this._boxKey);
+  MemoryCacheStorage(this._boxKey);
 
   @override
   Future<void> ensureInitialized() => Future.value();
 
-  /// Clears all data stored by all repositories in [SimpleMemoryCacheStorage].
+  /// Clears all data stored by all repositories in [MemoryCacheStorage].
   static Future<void> clearAll() async {
     _boxes.clear();
   }
 
-  Future<Map<K, CacheEntry<V>>> _ensureBox() => _lock.synchronized(() async {
-        Map? box = _boxes[_boxKey];
+  Future<BehaviorSubject<Map<K, CacheEntry<V>>>> _ensureBox() =>
+      _lock.synchronized(() async {
+        BehaviorSubject<Map>? box = _boxes[_boxKey];
         if (box == null) {
-          box = <K, CacheEntry<V>>{};
+          box = BehaviorSubject.seeded(<K, CacheEntry<V>>{});
           _boxes[_boxKey] = box;
         }
-        return box as Map<K, CacheEntry<V>>;
+        return box as BehaviorSubject<Map<K, CacheEntry<V>>>;
       });
 
   @override
-  Future<void> clear() => _ensureBox().then((box) => box.clear());
+  Future<void> clear() => _ensureBox().then((box) => box.value = {});
 
   @override
   Future<CacheEntry<V>?> get(K cacheKey) async {
-    return (await _ensureBox())[cacheKey];
+    return (await _ensureBox()).value[cacheKey];
   }
 
   @override
   Future<void> put(K cacheKey, V data, {int? storeTime}) async {
     final box = await _ensureBox();
-    box[cacheKey] = CacheEntry(
-      data,
-      storeTime: storeTime ?? DateTime.now().millisecondsSinceEpoch,
-    );
+    box.value = {
+      ...box.value,
+      cacheKey: CacheEntry(
+        data,
+        storeTime: storeTime ?? DateTime.now().millisecondsSinceEpoch,
+      )
+    };
   }
 
   @override
   Future<void> delete(K cacheKey) async {
-    (await _ensureBox()).remove(cacheKey);
+    final box = await _ensureBox();
+    final newValue = {...box.value}..remove(cacheKey);
+    box.value = newValue;
   }
 
-  /// Not implemented yet
   @override
-  Stream<List<V>> watch() =>
-      throw UnsupportedError('Not supported yet for MemoryCacheStorage!');
+  Stream<List<V>> watch() => _ensureBox()
+      .asStream()
+      .switchMap((box) => box)
+      .map((map) => map.values.map((e) => e.data).toList());
 
   @override
   Future<List<V>> getAll() =>
-      _ensureBox().then((box) => box.values.map((e) => e.data).toList());
+      _ensureBox().then((box) => box.value.values.map((e) => e.data).toList());
 
   @override
   String toString() => 'MemoryCacheStorage($_boxKey)';