+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.3
+* SimpleMemoryCacheStorage renamed to MemoryCacheStorage
+* Implemented `watch()` for MemoryCacheStorage
+
 ## 0.0.2+2
 * Update readme